fix: move FrontMatterParser to parser package, cosmetic naming fixes
