+++ commonmark-ext-yaml-front-matter/src/main/java/module-info.java
@@ -1,5 +1,6 @@
 module org.commonmark.ext.front.matter {
     exports org.commonmark.ext.front.matter;
+    exports org.commonmark.ext.front.matter.parser;
 
     requires transitive org.commonmark;
 }
