+++ commonmark-ext-yaml-front-matter/src/main/java/org/commonmark/ext/front/matter/parser/FrontMatterParser.java
@@ -1,5 +1,6 @@
-package org.commonmark.ext.front.matter;
+package org.commonmark.ext.front.matter.parser;
 
+import org.commonmark.ext.front.matter.YamlFrontMatterBlock;
 import org.commonmark.parser.SourceLine;
 
 /**
