+++ commonmark-ext-yaml-front-matter/src/main/java/org/commonmark/ext/front/matter/parser/RawContentParser.java
@@ -1,6 +1,5 @@
 package org.commonmark.ext.front.matter.parser;
 
-import org.commonmark.ext.front.matter.FrontMatterParser;
 import org.commonmark.ext.front.matter.YamlFrontMatterBlock;
 import org.commonmark.ext.front.matter.YamlFrontMatterRawContent;
 import org.commonmark.parser.SourceLine;
